Fix pylint errors other than async-sync overrides
