+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_client.py
@@ -402,4 +402,3 @@ class ServiceBusClient(object):
             http_proxy=self._config.http_proxy,
             **kwargs
         )
-
