+++ sdk/servicebus/azure-servicebus/azure/servicebus/_servicebus_session_receiver.py
@@ -2,14 +2,9 @@
 # Copyright (c) Microsoft Corporation. All rights reserved.
 # Licensed under the MIT License. See License.txt in the project root for license information.
 # --------------------------------------------------------------------------------------------
-import time
 import logging
-from typing import Any, TYPE_CHECKING, Optional, Union
-import six
+from typing import Any, TYPE_CHECKING
 
-from ._base_handler import BaseHandler
-from ._common.utils import utc_from_timestamp, utc_now
-from ._common.constants import ReceiveSettleMode
 from ._common.receiver_mixins import SessionReceiverMixin
 from ._servicebus_receiver import ServiceBusReceiver
 from ._servicebus_session import ServiceBusSession
