+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_sender_async.py
@@ -265,7 +265,7 @@ class ServiceBusSender(BaseHandlerAsync, SenderMixin):
         """
         try:
             batch = await self.create_batch()
-            batch._from_list(message)
+            batch._from_list(message)  # pylint: disable=protected-access
             message = batch
         except TypeError:  # Message was not a list or generator.
             pass
