+++ sdk/servicebus/azure-servicebus/azure/servicebus/aio/_servicebus_session_receiver_async.py
@@ -5,8 +5,7 @@
 import logging
 from typing import Any, TYPE_CHECKING, List, Union
 
-from .._common.receiver_mixins import ReceiverMixin, SessionReceiverMixin
-from .._common.constants import ReceiveSettleMode
+from .._common.receiver_mixins import SessionReceiverMixin
 from ._servicebus_receiver_async import ServiceBusReceiver
 from ._servicebus_session_async import ServiceBusSession
