+++ sdk/servicebus/azure-servicebus/dev_requirements.txt
@@ -1,3 +1,4 @@
+-e ../../core/azure-core
 -e ../../../tools/azure-devtools
 -e ../../../tools/azure-sdk-tools
 -e ../azure-mgmt-servicebus
